# kanban-frontend/api.yaml
---
openapi: 3.1.0
info:
  title: FastAPI
  version: 0.1.0
paths:
  "/token":
    post:
      summary: Login
      operationId: login_token_post
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              "$ref": "#/components/schemas/Body_login_token_post"
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/signup":
    post:
      summary: Signup
      operationId: signup_signup_post
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/SignUpForm"
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/board/update":
    put:
      summary: Create Board
      operationId: create_board_board_update_put
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Board"
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
      security:
        - OAuth2PasswordBearer: []
  "/board/delete":
    delete:
      summary: Delete Board
      operationId: delete_board_board_delete_delete
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: board_id
          in: query
          required: true
          schema:
            title: Board Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/board/column/update":
    post:
      summary: Update Column
      operationId: update_column_board_column_update_post
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: board_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
            title: Board Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Column"
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/board/column/delete":
    delete:
      summary: Delete Column
      operationId: delete_column_board_column_delete_delete
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: board_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
            title: Board Id
        - name: colum_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
            title: Colum Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/task/move":
    post:
      summary: Move Task
      operationId: move_task_task_move_post
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: task_id
          in: query
          required: true
          schema:
            type: string
            title: Task Id
        - name: dest
          in: query
          required: true
          schema:
            type: string
            format: uuid
            title: Dest
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/task/update":
    put:
      summary: Create Task
      operationId: create_task_task_update_put
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/Task"
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
      security:
        - OAuth2PasswordBearer: []
  "/task/delete":
    delete:
      summary: Create Task
      operationId: create_task_task_delete_delete
      security:
        - OAuth2PasswordBearer: []
      parameters:
        - name: task_id
          in: query
          required: true
          schema:
            title: Task Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/HTTPValidationError"
  "/":
    get:
      summary: Root
      operationId: root__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
  "/auth_debug":
    get:
      summary: Root
      operationId: root_auth_debug_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
        - OAuth2PasswordBearer: []
components:
  schemas:
    Board:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        members:
          items:
            type: string
          type: array
          title: Members
        columns:
          items:
            type: string
            format: uuid
          type: array
          title: Columns
      type: object
      required:
        - name
        - members
        - columns
      title: Board
    Body_login_token_post:
      properties:
        grant_type:
          anyOf:
            - type: string
              pattern: password
            - type: 'null'
          title: Grant Type
        username:
          type: string
          title: Username
        password:
          type: string
          title: Password
        scope:
          type: string
          title: Scope
          default: ''
        client_id:
          anyOf:
            - type: string
            - type: 'null'
          title: Client Id
        client_secret:
          anyOf:
            - type: string
            - type: 'null'
          title: Client Secret
      type: object
      required:
        - username
        - password
      title: Body_login_token_post
    Column:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
      type: object
      required:
        - name
      title: Column
    HTTPValidationError:
      properties:
        detail:
          items:
            "$ref": "#/components/schemas/ValidationError"
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    SignUpForm:
      properties:
        username:
          type: string
          title: Username
        password:
          type: string
          title: Password
      type: object
      required:
        - username
        - password
      title: SignUpForm
    Task:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        board_id:
          type: string
          format: uuid
          title: Board Id
        title:
          type: string
          title: Title
        assignee:
          type: string
          title: Assignee
        assigned:
          items:
            type: string
          type: array
          title: Assigned
        column_id:
          type: string
          title: Column Id
          default: Todos
      type: object
      required:
        - board_id
        - title
        - assignee
        - assigned
      title: Task
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
        - loc
        - msg
        - type
      title: ValidationError
  securitySchemes:
    OAuth2PasswordBearer:
      type: oauth2
      flows:
        password:
          scopes: {}
          tokenUrl: token