# kanban-frontend/gen/src/model/Board.js
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 *
 */

import ApiClient from '../ApiClient';

/**
 * The Board model module.
 * @module model/Board
 * @version 0.1.0
 */
class Board {
    /**
     * Constructs a new <code>Board</code>.
     * @alias module:model/Board
     * @param name {Object} 
     * @param members {Object} 
     * @param columns {Object} 
     */
    constructor(name, members, columns) { 
        
        Board.initialize(this, name, members, columns);
    }

    /**
     * Initializes the fields of this object.
     * This method is used by the constructors of any subclasses, in order to implement multiple inheritance (mix-ins).
     * Only for internal use.
     */
    static initialize(obj, name, members, columns) { 
        obj['name'] = name;
        obj['members'] = members;
        obj['columns'] = columns;
    }

    /**
     * Constructs a <code>Board</code> from a plain JavaScript object, optionally creating a new instance.
     * Copies all relevant properties from <code>data</code> to <code>obj</code> if supplied or a new instance if not.
     * @param {Object} data The plain JavaScript object bearing properties of interest.
     * @param {module:model/Board} obj Optional instance to populate.
     * @return {module:model/Board} The populated <code>Board</code> instance.
     */
    static constructFromObject(data, obj) {
        if (data) {
            obj = obj || new Board();

            if (data.hasOwnProperty('id')) {
                obj['id'] = ApiClient.convertToType(data['id'], Object);
            }
            if (data.hasOwnProperty('name')) {
                obj['name'] = ApiClient.convertToType(data['name'], Object);
            }
            if (data.hasOwnProperty('members')) {
                obj['members'] = ApiClient.convertToType(data['members'], Object);
            }
            if (data.hasOwnProperty('columns')) {
                obj['columns'] = ApiClient.convertToType(data['columns'], Object);
            }
        }
        return obj;
    }


}

/**
 * @member {Object} id
 */
Board.prototype['id'] = undefined;

/**
 * @member {Object} name
 */
Board.prototype['name'] = undefined;

/**
 * @member {Object} members
 */
Board.prototype['members'] = undefined;

/**
 * @member {Object} columns
 */
Board.prototype['columns'] = undefined;






export default Board;