# kanban-frontend/gen/src/model/Column.js
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 *
 */

import ApiClient from '../ApiClient';

/**
 * The Column model module.
 * @module model/Column
 * @version 0.1.0
 */
class Column {
    /**
     * Constructs a new <code>Column</code>.
     * @alias module:model/Column
     * @param name {Object} 
     */
    constructor(name) { 
        
        Column.initialize(this, name);
    }

    /**
     * Initializes the fields of this object.
     * This method is used by the constructors of any subclasses, in order to implement multiple inheritance (mix-ins).
     * Only for internal use.
     */
    static initialize(obj, name) { 
        obj['name'] = name;
    }

    /**
     * Constructs a <code>Column</code> from a plain JavaScript object, optionally creating a new instance.
     * Copies all relevant properties from <code>data</code> to <code>obj</code> if supplied or a new instance if not.
     * @param {Object} data The plain JavaScript object bearing properties of interest.
     * @param {module:model/Column} obj Optional instance to populate.
     * @return {module:model/Column} The populated <code>Column</code> instance.
     */
    static constructFromObject(data, obj) {
        if (data) {
            obj = obj || new Column();

            if (data.hasOwnProperty('id')) {
                obj['id'] = ApiClient.convertToType(data['id'], Object);
            }
            if (data.hasOwnProperty('name')) {
                obj['name'] = ApiClient.convertToType(data['name'], Object);
            }
        }
        return obj;
    }


}

/**
 * @member {Object} id
 */
Column.prototype['id'] = undefined;

/**
 * @member {Object} name
 */
Column.prototype['name'] = undefined;






export default Column;