# kanban-frontend/gen/src/model/SignUpForm.js
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 *
 */

import ApiClient from '../ApiClient';

/**
 * The SignUpForm model module.
 * @module model/SignUpForm
 * @version 0.1.0
 */
class SignUpForm {
    /**
     * Constructs a new <code>SignUpForm</code>.
     * @alias module:model/SignUpForm
     * @param username {Object} 
     * @param password {Object} 
     */
    constructor(username, password) { 
        
        SignUpForm.initialize(this, username, password);
    }

    /**
     * Initializes the fields of this object.
     * This method is used by the constructors of any subclasses, in order to implement multiple inheritance (mix-ins).
     * Only for internal use.
     */
    static initialize(obj, username, password) { 
        obj['username'] = username;
        obj['password'] = password;
    }

    /**
     * Constructs a <code>SignUpForm</code> from a plain JavaScript object, optionally creating a new instance.
     * Copies all relevant properties from <code>data</code> to <code>obj</code> if supplied or a new instance if not.
     * @param {Object} data The plain JavaScript object bearing properties of interest.
     * @param {module:model/SignUpForm} obj Optional instance to populate.
     * @return {module:model/SignUpForm} The populated <code>SignUpForm</code> instance.
     */
    static constructFromObject(data, obj) {
        if (data) {
            obj = obj || new SignUpForm();

            if (data.hasOwnProperty('username')) {
                obj['username'] = ApiClient.convertToType(data['username'], Object);
            }
            if (data.hasOwnProperty('password')) {
                obj['password'] = ApiClient.convertToType(data['password'], Object);
            }
        }
        return obj;
    }


}

/**
 * @member {Object} username
 */
SignUpForm.prototype['username'] = undefined;

/**
 * @member {Object} password
 */
SignUpForm.prototype['password'] = undefined;






export default SignUpForm;